Font Library: Test frontend and classic themes
